+++ src/editor/windows/save_map.rs
@@ -10,6 +10,7 @@ use crate::{
     Resources,
 };
 
+#[derive(Default)]
 pub struct SaveMapWindow {
     map_name: String,
     overwrite_existing: bool,
